Move SyncRefinementFlags to its own header
I ended up reverting the use of this class in other .C files, but I
might need it elsewhere eventually, and even if that doesn't happen,
it feels better-organized to put this class in its own file.
