+++ include/libmesh/Makefile.am
@@ -195,6 +195,7 @@ BUILT_SOURCES = \
         postscript_io.h \
         replicated_mesh.h \
         serial_mesh.h \
+        sync_refinement_flags.h \
         tecplot_io.h \
         tetgen_io.h \
         ucd_io.h \
@@ -1094,6 +1095,9 @@ replicated_mesh.h: $(top_srcdir)/include/mesh/replicated_mesh.h
 	$(AM_V_GEN)rm -f $@ && $(LN_S) -f $< $@
 
 serial_mesh.h: $(top_srcdir)/include/mesh/serial_mesh.h
+	$(AM_V_GEN)rm -f $@ && $(LN_S) -f $< $@
+
+sync_refinement_flags.h: $(top_srcdir)/include/mesh/sync_refinement_flags.h
 	$(AM_V_GEN)rm -f $@ && $(LN_S) -f $< $@
 
 tecplot_io.h: $(top_srcdir)/include/mesh/tecplot_io.h
